requirements.txt
pymongo
pylint
python-dateutil
wmi